Struct inexor_rgf_reactive_model_impl::relations::reactive_relation::ReactiveRelation

pub struct ReactiveRelation(/* private fields */);

Implementations

impl ReactiveRelation

pub fn builder() -> ReactiveRelationInstanceBuilder

pub fn builder_with_entities( outbound: ReactiveEntity, ty: &RelationInstanceTypeId, inbound: ReactiveEntity ) -> ReactiveRelationInstanceBuilder<((ReactiveEntity,), (RelationInstanceTypeId,), (ReactiveEntity,), (), (), (), ())>

pub fn builder_with_entities_and_properties( outbound: ReactiveEntity, ty: &RelationInstanceTypeId, inbound: ReactiveEntity, properties: &PropertyTypes ) -> ReactiveRelationInstanceBuilder<((ReactiveEntity,), (RelationInstanceTypeId,), (ReactiveEntity,), (), (ReactiveProperties<RelationInstanceId>,), (), ())>

Creates a builder for the given relation instance type id. Generates an id for the reactive relation. Converts property types into reactive properties and initializes the properties with default values.

pub fn builder_from_type_with_unique_id( outbound: ReactiveEntity, relation_type: &RelationType, inbound: ReactiveEntity ) -> ReactiveRelationInstanceBuilder<((ReactiveEntity,), (RelationInstanceTypeId,), (ReactiveEntity,), (), (ReactiveProperties<RelationInstanceId>,), (), ())>

pub fn builder_from_type_with_unique_instance_id( outbound: ReactiveEntity, relation_type: &RelationType, instance_id: String, inbound: ReactiveEntity ) -> ReactiveRelationInstanceBuilder<((ReactiveEntity,), (RelationInstanceTypeId,), (ReactiveEntity,), (), (ReactiveProperties<RelationInstanceId>,), (), ())>

pub fn builder_from_type_with_random_instance_id( outbound: ReactiveEntity, relation_type: &RelationType, inbound: ReactiveEntity ) -> ReactiveRelationInstanceBuilder<((ReactiveEntity,), (RelationInstanceTypeId,), (ReactiveEntity,), (), (ReactiveProperties<RelationInstanceId>,), (), ())>

impl ReactiveRelation

pub fn new_from_properties( outbound: ReactiveEntity, ty: RelationInstanceTypeId, inbound: ReactiveEntity, properties: DashMap<String, Value> ) -> ReactiveRelation

pub fn new_from_instance( outbound: ReactiveEntity, inbound: ReactiveEntity, instance: RelationInstance ) -> ReactiveRelation